add sms collection to mongoConfig
