+++ app/mongo/collections/index.js
@@ -1,5 +1,7 @@
 import emailsCollection from './emails.js'
+import smsCollection from './sms.js'
 
 export {
-  emailsCollection
+  emailsCollection,
+  smsCollection
 }
